lindorm-vdb finished, tests pass
